requirements.txt
Flask
requests
python-dotenv
PyJWT

Flask-Bcrypt
Flask-SQLAlchemy
Flask-Cors